fix: refacto trainer
