+++ src/steps/model_training/paddle_ocr_trainer.py
@@ -9,8 +9,8 @@ from src.models.parameters.common.paddle_ocr.paddle_ocr_augmentation_parameters 
 from src.models.parameters.common.paddle_ocr.paddle_ocr_hyper_parameters import (
     PaddleOCRHyperParameters,
 )
-from src.models.steps.model_training.paddle_ocr_model_trainer import (
-    PaddleOCRModelTrainer,
+from src.models.steps.model_training.paddle_ocr_model_collection_trainer import (
+    PaddleOCRModelCollectionTrainer,
 )
 
 
@@ -19,10 +19,10 @@ def paddle_ocr_trainer(model_collection):
     context: PicselliaTrainingContext[
         PaddleOCRHyperParameters, PaddleOCRAugmentationParameters
     ] = Pipeline.get_active_context()
-    model_trainer = PaddleOCRModelTrainer(
+    model_trainer = PaddleOCRModelCollectionTrainer(
         model_collection=model_collection, experiment=context.experiment
     )
-    model_collection = model_trainer.train(
+    model_collection = model_trainer.train_model_collection(
         bbox_epochs=context.hyperparameters.bbox_epochs,
         text_epochs=context.hyperparameters.text_epochs,
     )
